requirements.txt
coverage
Django
django-debug-toolbar
django-filter
djangorestframework
drf-yasg
factory-boy
Faker
ipython
psycopg2-binary

